+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## CHANGELOG SO FAR...
 
+### *[09/09/2021] 1.0.0-alpha.3*
+* `[+] Added option {removeSourceMappingTag: boolean}, to remove //#sourceMappingURL=... tag`
+* `[b] Apply srcNode mapping AFTER the transform option is run, if any`
+* `[b] Normalize file endlines on mocha tests, for consistency`
+
 ### *[20/05/2021] 1.0.0-alpha.2*
 * `[=] Readme enhancement`